+++ src/test/java/de/chrgroth/jsonstore/store/JsonSingletonStoreTest.java
@@ -1,4 +1,4 @@
-package com.github.christiangroth.jsonstore.store;
+package de.chrgroth.jsonstore.store;
 
 import java.io.File;
 import java.nio.charset.StandardCharsets;
@@ -8,6 +8,9 @@ import org.junit.Before;
 import org.junit.Test;
 
 import com.google.common.io.Files;
+
+import de.chrgroth.jsonstore.json.FlexjsonHelperTest;
+import de.chrgroth.jsonstore.store.JsonSingletonStore;
 
 public class JsonSingletonStoreTest {
 	
@@ -22,10 +25,10 @@ public class JsonSingletonStoreTest {
 	@Before
 	public void init() {
 		tempDir = Files.createTempDir();
-		persistentStore = new JsonSingletonStore<>(String.class, tempDir, StandardCharsets.UTF_8, true, true);
-		persistentStoreCopy = new JsonSingletonStore<>(String.class, tempDir, StandardCharsets.UTF_8, true, true);
-		transientStore = new JsonSingletonStore<>(String.class, null, null, false, false);
-		transientStoreCopy = new JsonSingletonStore<>(String.class, null, null, false, false);
+		persistentStore = new JsonSingletonStore<>(String.class, FlexjsonHelperTest.DATE_TIME_PATTERN, tempDir, StandardCharsets.UTF_8, true, true);
+		persistentStoreCopy = new JsonSingletonStore<>(String.class, FlexjsonHelperTest.DATE_TIME_PATTERN, tempDir, StandardCharsets.UTF_8, true, true);
+		transientStore = new JsonSingletonStore<>(String.class, FlexjsonHelperTest.DATE_TIME_PATTERN, null, null, false, false);
+		transientStoreCopy = new JsonSingletonStore<>(String.class, FlexjsonHelperTest.DATE_TIME_PATTERN, null, null, false, false);
 		testDataOne = "test data";
 	}